Fixing shop keeps for Fey'ri as well

+++ d/darkwood/camps/mon/caldur.c
@@ -30,17 +30,32 @@ void init() {
    }
 }
 
-void kill_evil() {
-   string race;
-   race = TP->query_race();
-      if( race == "drow" || race == "half-drow" || race == "goblin" || race == "hobgoblin" ||
-      race == "orc" || race == "half-orc" || race == "ogre" || race == "half-ogre" ||
-      race == "kobold" || race == "ogre-mage" || race == "bugbear" || race == "wemic" ||
-      race == "gnoll") {
-   force_me("yell %^BOLD%^%^GREEN%^Someone help me!%^RESET%^");
-   force_me("kill "+TP->query_name());
-   return;
-   }
+void kill_evil()
+{
+    string race;
+    string subrace;
+    object shape;
+    race = TP->query_race();
+    subrace = TP->query("subrace");
+    if (objectp(shape = TP->query_property("shapeshifted"))) {
+        race = (string)shape->query_shape_race();
+    }
+    if (objectp(shape = TP->query_property("altered"))) {
+        race = (string)shape->query_shape_race();
+        subrace = race;
+    }
+    if (race == "drow" || race == "half-drow" || race == "goblin" || race == "hobgoblin" ||
+        race == "orc" || race == "half-orc" || race == "ogre" || race == "half-ogre" ||
+        race == "kobold" || race == "ogre-mage" || race == "bugbear" || race == "wemic" ||
+        race == "gnoll" ||
+        race == "troll" ||
+        race == "ratkin" ||
+        race == "wererat" ||
+        subrace == "fey'ri" ) {
+        force_me("yell %^BOLD%^%^GREEN%^Someone help me!%^RESET%^");
+        force_me("kill " + TP->query_name());
+        return;
+    }
 }
